Add more enviromental config
Add PACKETBEAT_LISTEN_PORTS, should be set to a YAML array of ports that
should be listened on.

Add PACKETBEAT_MAX_MESSAGE_SIZE, number in bytes of the largest message
for which the body will be stored.

Also add SPARQL content types to types for which the body will be stored.